+++ lib/cinch/bot.rb
@@ -284,8 +284,8 @@ module Cinch
           wait = 2**@reconnects
           wait = @config.max_reconnect_delay if wait > @config.max_reconnect_delay
           @loggers.info "Waiting #{wait} seconds before reconnecting"
-          start_time = Time.now
-          while !@quitting && (Time.now - start_time) < wait
+          start_time = Process.clock_gettime(Process::CLOCK_MONOTONIC)
+          while !@quitting && (Process.clock_gettime(Process::CLOCK_MONOTONIC) - start_time) < wait
             sleep 1
           end
         end